fix: refactor connector messaging destinations auto configuration
