+++ activiti-cloud-runtime-bundle-service/activiti-cloud-services-runtime-bundle/activiti-cloud-services-connectors/src/main/java/org/activiti/services/connectors/conf/CloudConnectorsAutoConfiguration.java
@@ -20,7 +20,6 @@ import org.activiti.cloud.services.events.configuration.RuntimeBundleProperties;
 import org.activiti.cloud.services.events.converter.RuntimeBundleInfoAppender;
 import org.activiti.cloud.services.events.listeners.ProcessEngineEventsAggregator;
 import org.activiti.engine.ManagementService;
-import org.activiti.engine.RepositoryService;
 import org.activiti.engine.RuntimeService;
 import org.activiti.engine.impl.bpmn.parser.factory.DefaultActivityBehaviorFactory;
 import org.activiti.engine.impl.persistence.entity.integration.IntegrationContextManager;
@@ -136,25 +135,4 @@ public class CloudConnectorsAutoConfiguration {
         taskBehavior.setVariablesCalculator(variablesMappingProvider);
         return taskBehavior;
     }
-
-    @Bean
-    @ConditionalOnMissingBean
-    public ConnectorImplementationsProvider connectorDestinationsProvider(RepositoryService repositoryService) {
-        return new RepositoryConnectorImplementationsProvider(repositoryService);
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    public ConnectorDestinationMappingStrategy destinationMappingStrategy() {
-        return new ConnectorDestinationMappingStrategy() {};
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    public ConnectorDestinationsBeanPostProcessor connectorDestinationsBeanPostProcessor(ConnectorImplementationsProvider destinationsProvider,
-                                                                                         ConnectorDestinationMappingStrategy destinationMappingStrategy) {
-        return new ConnectorDestinationsBeanPostProcessor(destinationsProvider,
-                                                          destinationMappingStrategy);
-    }
-
 }
