+++ activiti-cloud-runtime-bundle-service/activiti-cloud-services-runtime-bundle/activiti-cloud-services-connectors/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-    org.activiti.services.connectors.conf.CloudConnectorsAutoConfiguration
+    org.activiti.services.connectors.conf.CloudConnectorsAutoConfiguration,\
+    org.activiti.services.connectors.conf.CloudConnectorsMessagingAutoConfiguration
